LayoutTests/imported/w3c/web-platform-tests/css/css-align/blocks/justify-self-vlr-ltr-htb-expected.txt

















PASS .item 1
FAIL .item 2 assert_equals:
<div class="container">
  <div class="item" style="justify-self: center;" data-expected-height="20" data-offset-y="10"></div>
</div>
offsetTop expected 10 but got 0
FAIL .item 3 assert_equals:
<div class="container">
  <div class="item" style="justify-self: end;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
PASS .item 4
FAIL .item 5 assert_equals:
<div class="container">
  <div class="item" style="justify-self: self-end;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
PASS .item 6
FAIL .item 7 assert_equals:
<div class="container">
  <div class="item" style="justify-self: right;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
FAIL .item 8 assert_equals:
<div class="container">
  <div class="item" style="justify-self: stretch;" data-expected-height="40" data-offset-y="0"></div>
</div>
height expected 40 but got 20
PASS .item 9
FAIL .item 10 assert_equals:
<div class="container">
  <div class="item rtl" style="justify-self: center;" data-expected-height="20" data-offset-y="10"></div>
</div>
offsetTop expected 10 but got 0
FAIL .item 11 assert_equals:
<div class="container">
  <div class="item rtl" style="justify-self: end;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
PASS .item 12
FAIL .item 13 assert_equals:
<div class="container">
  <div class="item rtl" style="justify-self: self-end;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
PASS .item 14
FAIL .item 15 assert_equals:
<div class="container">
  <div class="item rtl" style="justify-self: right;" data-expected-height="20" data-offset-y="20"></div>
</div>
offsetTop expected 20 but got 0
FAIL .item 16 assert_equals:
<div class="container">
  <div class="item rtl" style="justify-self: stretch;" data-expected-height="40" data-offset-y="0"></div>
</div>
height expected 40 but got 20
